requirements.txt
# 后端所需的 Python 库

# Web框架
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.0
Werkzeug>=2.3.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0
pyecharts>=2.0.0

# 机器学习和AI
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0

# 文件处理
openpyxl>=3.1.0
xlrd>=2.0.0
Pillow>=10.0.0

# 模板引擎
Jinja2>=3.1.0

# HTTP请求
requests>=2.31.0
urllib3>=2.0.0

# 数据库和缓存
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0
redis>=4.6.0

# 配置和环境
python-dotenv>=1.0.0
click>=8.1.0

# 日志和监控
loguru>=0.7.0

# 数据验证
pydantic>=2.0.0
marshmallow>=3.19.0

# 异步支持
aiohttp>=3.8.0
aiofiles>=23.0.0

# 工具库
tqdm>=4.65.0
more-itertools>=9.1.0
psutil>=5.9.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 代码质量
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# 安全
cryptography>=41.0.0

# 时间处理
python-dateutil>=2.8.0

# 缓存
redis>=4.6.0

# 任务队列
celery>=5.3.0

# 文档生成
sphinx>=7.0.0